# api/openapi.yaml
openapi: "3.1.0"
info:
  title: rssembly API
  version: "0.1.0"
  description: |
    rssembly is an open-source, self-hostable RSS reader platform.
    One backend, multiple clients — all with privacy in mind.
servers:
  - url: http://localhost:8080
    description: Local development
paths:
...
  /api/v1/articles/{articleID}/read-state:
    put:
      operationId: setReadState
      summary: Mark an article as read/unread/saved
      tags: [Articles]
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      parameters:
        - name: articleID
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [state]
              properties:
                state:
                  type: string
                  enum: [unread, read, saved]
      responses:
        "200":
          description: State updated
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...